docs: add xml-docs for basic models
