+++ src/HardCoded.MockServer.Contracts/Models/ValueTypes/Delay.cs
@@ -1,10 +1,16 @@
-using HardCoded.MockServer.Contracts.Abstractions;
+using System;
 
-namespace HardCoded.MockServer.Models.ValueTypes
+using HardCoded.MockServer.Contracts.Abstractions;
+using HardCoded.MockServer.Contracts.Models.HttpEntities;
+
+
+namespace HardCoded.MockServer.Contracts.Models.ValueTypes
 {
-   
-    public class Delay  : BuildableBase
-    {
+   /// <summary>
+   /// Model to configure an optional <see cref="Delay"/> before responding with an action to a matched <see cref="HttpRequest"/>.
+   /// </summary>
+    public class Delay  : BuildableBase, IEquatable<Delay>
+   {
         public Delay()
         {
         }
@@ -19,8 +25,14 @@ namespace HardCoded.MockServer.Models.ValueTypes
         public static Delay FromMinutes(int minutes) => new Delay(minutes, TimeUnit.MINUTES);
         public static Delay None => new Delay();
         
+        /// <summary>
+        /// The <see cref="TimeUnit"/> of the <see cref="Delay"/>.
+        /// </summary>
         public TimeUnit TimeUnit { get; set; }
 
+        /// <summary>
+        /// The value of the <see cref="Delay"/>.
+        /// </summary>
         public int Value { get; set; }
     }
 }
