--- src/HardCoded.MockServer/Models/HttpEntities/HttpForward.cs
@@ -1,12 +1,0 @@
-using HardCoded.MockServer.Contracts.Abstractions;
-
-
-namespace HardCoded.MockServer.Models.HttpEntities
-{
-    public class HttpForward : BuildableBase
-    {
-        public string Host { get; set; }
-        
-        public long Port { get; set; }
-    }
-}
